+++ src/trans2.h
@@ -226,9 +226,13 @@ Byte offset   Type     name                description
 #define NT_FILE_ATTRIBUTE_NORMAL 0x80
 
 void mask_convert(char *mask);
-int reply_findclose(char *inbuf, char *outbuf, int inbuf_len, int outbuf_len);
-int reply_findnclose(char *inbuf, char *outbuf, int inbuf_len, int outbuf_len);
-int reply_transs2(char *inbuf, char *outbuf, int inbuf_len, int outbuf_len);
-int reply_trans2(char *inbuf, char *outbuf, int inbuf_len, int outbuf_len);
+int reply_findclose(char *inbuf, char *outbuf, size_t inbuf_len,
+                    size_t outbuf_len);
+int reply_findnclose(char *inbuf, char *outbuf, size_t inbuf_len,
+                     size_t outbuf_len);
+int reply_transs2(char *inbuf, char *outbuf, size_t inbuf_len,
+                  size_t outbuf_len);
+int reply_trans2(char *inbuf, char *outbuf, size_t inbuf_len,
+                 size_t outbuf_len);
 
 #endif
